+++ CHANGELOG.md
@@ -7,3 +7,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 - Built initial Adapter interface
 - Implemented PdfLayer adapter
+- Implemented Doppio adapter
+- Restructured with Request interface